sight/requirements.txt
absl-py==1.4.0
ale-py==0.7.5
asttokens==2.2.1
astunparse==1.6.3
atari-py==0.2.9
attrs==22.1.0
avro==1.11.3
backcall==0.2.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.82
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.7
cryptography==41.0.6
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
distrax==0.1.4
# Editable install with no version control (dm-acme==0.4.1)
# -e /usr/local/google/home/meetashah/envs/sight_310/lib/python3.10/site-packages/acme-latest-github
git+https://github.com/deepmind/acme@master
dm-env==1.6
dm-haiku==0.0.9
dm-launchpad==0.5.2
dm-reverb==0.7.2
dm-sonnet==2.0.1
dm-tree==0.1.8
docstring-parser==0.15
dsub==0.4.7
executing==1.2.0
fastavro==1.7.0
filelock==3.8.0
fire==0.4.0
flatbuffers==1.12
fonttools==4.38.0
funcsigs==1.0.2
gast==0.4.0
glfw==2.5.6
google-api-core==1.33.2
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.18.3
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.5
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-cloud-notebooks==1.4.4
google-cloud-pipeline-components==1.0.26
google-cloud-resource-manager==1.6.3
google-cloud-spanner==3.26.0
google-cloud-storage==1.44.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-gcp==0.2.2
grpcio-status==1.48.2
grpcio-tools==1.48.2
gym==0.25.0
gym-notices==0.0.8
h5py==3.7.0
httplib2==0.20.4
idna==3.4
imageio==2.25.1
importlib-resources==6.0.0
ipython==8.14.0
jax==0.4.13
jaxlib==0.4.13
jedi==0.18.2
jmp==0.0.2
joblib==1.2.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kfp==1.8.14
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kiwisolver==1.4.4
kubernetes==18.20.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mock==4.0.3
ndg-httpsclient==0.5.1
numproto==0.3.0
numpy==1.25.1
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.5
overrides==7.3.1
packaging==21.3
pandas==1.5.1
parameterized==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==2.5.3
portpicker==1.5.2
prompt-toolkit==3.0.39
proto-plus==1.22.1
protobuf==3.20.2
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.15.1
PyOpenGL==3.1.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
PyYAML==5.4.1
readerwriterlock==1.0.9
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rlax==0.1.6
rsa==4.8
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.4
stack-data==0.6.2
strip-hints==0.1.10
tabulate==0.8.9
tenacity==7.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.15.0
termcolor==2.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
toolz==0.12.0
traitlets==5.9.0
trfl==1.2.0
typer==0.6.1
typing_extensions==4.5.0
uritemplate==3.0.1
urllib3==1.26.18
virtualenv==20.16.6
wcwidth==0.2.6
websocket-client==1.4.1
Werkzeug==2.3.8
wrapt==1.14.1

